+++ tests/profiles/fontwerk_test.py
@@ -25,8 +25,7 @@ def test_check_vendor_id():
     check = CheckTester(fontwerk_profile,
                         'com.fontwerk/check/vendor_id')
 
-    font = TEST_FILE('abeezee/ABeeZee-Italic.ttf')
-    ttFont = TTFont(font)
+    ttFont = TTFont(TEST_FILE('abeezee/ABeeZee-Italic.ttf'))
     assert_results_contain(check(ttFont),
                            FAIL, 'bad-vendor-id',
                            ", but should be 'WERK'.")
@@ -40,8 +39,7 @@ def test_check_weight_class_fvar():
     check = CheckTester(fontwerk_profile,
                         'com.fontwerk/check/weight_class_fvar')
 
-    font = TEST_FILE('varfont/Oswald-VF.ttf')
-    ttFont = TTFont(font)
+    ttFont = TTFont(TEST_FILE('varfont/Oswald-VF.ttf'))
     assert_PASS(check(ttFont),
                 "matches fvar default value.")
 
@@ -66,3 +64,20 @@ def test_check_inconsistencies_between_fvar_stat():
     assert_results_contain(check(ttFont),
                            FAIL, 'missing-fvar-instance-axis-value',
                            'missing in STAT table')
+
+
+def test_check_style_linking():
+    check = CheckTester(fontwerk_profile,
+                        'com.fontwerk/check/style_linking')
+
+    font = TEST_FILE("bad_fonts/style_linking_issues/NotoSans-BoldItalic.ttf")
+    assert_results_contain(check(font),
+                           FAIL, 'style-linking-issue')
+
+    font = TEST_FILE("bad_fonts/style_linking_issues/NotoSans-Bold.ttf")
+    assert_results_contain(check(font),
+                           FAIL, 'style-linking-issue')
+
+    font = TEST_FILE("bad_fonts/style_linking_issues/NotoSans-MediumItalic.ttf")
+    assert_PASS(check(font),
+                "Style linking looks good.")